FIX: mapbox fixed but not really
