+++ app/assets/stylesheets/application.scss
@@ -9,7 +9,8 @@
 @import "font-awesome";
 @import "simple_calendar";
 @import "flatpickr/dist/flatpickr";
-@import 'select2/dist/css/select2';
+@import "select2/dist/css/select2";
+@import "mapbox-gl/dist/mapbox-gl";
 
 // Your CSS partials
 @import "components/index";
