--- app/javascript/packs/map.js
@@ -1,24 +1,0 @@
-import 'mapbox-gl/dist/mapbox-gl.css'
-import mapboxgl from 'mapbox-gl/dist/mapbox-gl.js';
-
-const mapElement = document.getElementById('map');
-
-if (mapElement) {
-  mapboxgl.accessToken = process.env.MAPBOX_API_KEY;
-
-  const map = new mapboxgl.Map({
-    container: 'map',
-    style: 'mapbox://styles/mapbox/streets-v10'
-  });
-
-  const markers = JSON.parse(mapElement.dataset.markers);
-
-    markers.forEach((marker) => {
-      new mapboxgl.Marker()
-        .setLngLat([marker.lng, marker.lat])
-        .addTo(map);
-    })
-
-  map.setZoom(8);
-  map.setCenter([markers[0].lng, markers[0].lat]);
-}
