+++ app/views/events/show.html.erb
@@ -27,7 +27,7 @@
           style="width: 100%;
           height: 300px;"
           data-markers="<%= @marker.to_json %>"
-        ></div>
+          data-mapbox-api-key="<%= ENV['MAPBOX_API_KEY'] %>"></div>
       </div>
 
       <div>
